% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axx.R
\name{axx}
\alias{axx}
\title{Adds axis in preferred default style}
\usage{
axx(x = T, y = T, box = T, cex.axis = 1.3)
}
\arguments{
\item{x}{Add the x-axis. Defaults to TRUE}

\item{y}{Add the y-axis. Defaults to TRUE}

\item{box}{Add axis lines. Defaults to TRUE}

\item{cex.axis}{Sets axis font size. Defaults to 1.3}
}
\description{
This function prepares the plotting area with a blank plot
}
\examples{
axx()
axx(X=FALSE) # Adds y-axis only.
}
\keyword{plot}
