% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blankplot.R
\name{blankplot}
\alias{blankplot}
\title{prepare plotting area with a blank plot}
\usage{
blankplot(x = 0:1, y = 0:1, ...)
}
\arguments{
\item{x}{X values. Defaults to 0:1}

\item{y}{Y values. Defaults to 0:1}
}
\description{
This function prepares the plotting area with a blank plot
}
\examples{
blankplot(1:10, 1:10)

}
\keyword{plot}
