% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smalUtilityFunctions.R
\name{checkCVarNames}
\alias{checkCVarNames}
\title{Checks if the names of a character vector or colnames of a data frames are valid C varibles}
\usage{
checkCVarNames(x)
}
\arguments{
\item{x}{data frame or caracter vectror}
}
\value{
a character vector with names who are not valid C varibles, if it returns
character(0) all names are fine
}
\description{
Checks if the names of a character vector or colnames of a data frames are valid C varibles
}
\author{
Johan Gudmundsson, \email{jgu@blackwoodseven.com}
}

