% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smalUtilityFunctions.R
\name{checkIfVarinaceExcist}
\alias{checkIfVarinaceExcist}
\title{Checks if variance excits for selected columns in a data frame, and retuns the column names where variance excist}
\usage{
checkIfVarinaceExcist(names, df, startDate)
}
\arguments{
\item{names}{names of columns to be checked}

\item{df}{a data frame with a date column}

\item{startDate}{the starting date of where the evaluation should start}
}
\value{
a character vector
}
\description{
This function checks if variance excits for selected columns in a data frame, and retuns the column names where variance excist
}
\author{
Johan Gudmundsson, \email{jgu@blackwoodseven.com}
}

