% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleannames.R
\name{cleannames}
\alias{cleannames}
\title{remove spaces and special characters in header of a data.frame or data.table}
\usage{
cleannames(df)
}
\arguments{
\item{df}{data.frame or data.table (required)}
}
\description{
This function imports a csv file into a data.table object
}
\examples{
df <- data.frame("f f"=c(1,1), "G&G"=c(5,2))
df_cleaned <- cleannames(df)
}
\keyword{dataManagement}
