% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanvect.R
\name{cleanvector}
\alias{cleanvector}
\title{Remove spaces and special characters from a vectoy}
\usage{
cleanvector(vect, interiorSpaces2Underscore = FALSE)
}
\arguments{
\item{vect}{filename No default (required)}

\item{interiorSpaces2Underscore}{Relative (or absolute) path to data directory}
}
\description{
This function special characters from a vector and converts spaces to underscore if desired
}
\examples{
cleanvector(c(1, 2, " aja alk", "\% s "), interiorSpaces2Underscore = TRUE)

}
\keyword{data}
