% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationsInBmodelList.R
\name{correlationsInBmodelList}
\alias{correlationsInBmodelList}
\title{A function retuns a data.frame with the correlations in the BModelList for the selected kpi}
\usage{
correlationsInBmodelList(kpi, BModelList)
}
\arguments{
\item{kpi}{kpi which should be investigated}

\item{BModelList}{the fitted BModelList}
}
\value{
A data frame with the correlation coefficents in descending order
}
\description{
This function sorts the lower triangular part of the correlation matrix for the fitted varibels in the BModelList
}
\examples{
correlationsInBmodelList(kpi = 'transactions', BModelList = bModelList)
}
\author{
Johan Gudmundsson, \email{jgu@blackwoodseven.com}
}

