% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devpdf.r
\name{devpdf}
\alias{devpdf}
\title{Adds axis labels in preferred style}
\usage{
devpdf(
  file,
  wd = "./graphs",
  open = TRUE,
  overwrite = FALSE,
  png = FALSE,
  tf = NULL,
  meta = FALSE
)
}
\arguments{
\item{file}{Text string in quotes pdf (and png) filename}

\item{wd}{Text string in quotes -- path to graph directory (defaults to "./graphs")}

\item{open}{Logical. Open resulting pdf file? Default = TRUE}

\item{overwrite}{Logical. Overwrite pdf file? Default = FALSE}

\item{png}{Logical. Make png file for inclusion in other applications (i.e., html)}

\item{tf}{Collects current .r file and path for output (Requires rstudioapi)}

\item{meta}{Logical. Prompts for figure number, notes and line #'s for output to text metafile}
}
\description{
This function outputs active plot to a pdf file and (optionally) a png, including (optionally) metadata
}
\examples{
devpdf("plotoutput", wd="", open=TRUE, overwrite=TRUE, png=TRUE, meta=FALSE)
}
\keyword{pdf}
\keyword{plot}
\keyword{png}
