% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devWin.r
\name{devWin}
\alias{devWin}
\title{Create a plotting window across platforms}
\usage{
devWin(height = 7, width = 7, rows = 1, cols = 1, oma = c(1, 1, 1, 1))
}
\arguments{
\item{height}{Plot window height}

\item{width}{Plot window width}

\item{rows}{Row count of plots.}

\item{cols}{Column count of plots.}

\item{oma}{outer margins (bottom, left, top, right)}
}
\description{
This function prepares the plotting area with a blank plot
}
\examples{
devWin(6,14, 1, 2, oma=c(2,2,1,1))
plot(1,1, main="left plot")
plot(1,1, main="right plot")
mtext("both plots x", outer=TRUE, side=1, cex=1.8)
mtext("both plots y", outer=TRUE, side=2, cex=1.8)
}
\keyword{plot}
