% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractVariblesFromBModelList.R
\name{extractVariblesFromBModelList}
\alias{extractVariblesFromBModelList}
\title{A function to extract fitted varibles from a BModelList}
\usage{
extractVariblesFromBModelList(kpi, BModelList)
}
\arguments{
\item{kpi}{The kpi which the varibles should be extracted for}

\item{BModelList}{The BModelList}
}
\value{
A data frame with extracted varibels
}
\description{
This function exctracts all the fitted varibles from the BModelList
}
\examples{
extractVariblesFromBModelList(kpi, BModelList)
}
\author{
Johan Gudmundsson, \email{jgu@blackwoodseven.com}
}

