% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getESS.R
\name{getESS}
\alias{getESS}
\title{A function to calculate the ESS of a STAN_fit}
\usage{
getESS(min_ess = 400, STAN_fit = NULL)
}
\arguments{
\item{min_ess}{the lowest ESS to not be returned}

\item{STAN_fit}{a stan fit df}
}
\value{
A data frame with the ESS below the min_ess limit in ascending order
}
\description{
This function returns the ESS of a STAN_fit ascending order
}
\examples{
getESS( min_ess = 900, STAN_fit = sfitdf)
}

