% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRHat.R
\name{getRHat}
\alias{getRHat}
\title{A function to calculate the RHat of a STAN_fit}
\usage{
getRHat(STAN_fit)
}
\arguments{
\item{STAN_fit}{a stan fit df}
}
\value{
A data frame with the RHat decending order
}
\description{
This function returns the RHat of a stan fit
}
\examples{
getRHat(STAN_fit = sfitdf)
}

