% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxx.R
\name{mtxx}
\alias{mtxx}
\title{Adds axis labels in preferred style}
\usage{
mtxx(x, y, line1 = 3, line2 = 3, cex = 1.4)
}
\arguments{
\item{x}{Text string in quotes for x-axis}

\item{y}{Text string in quotes for x-axis}

\item{line1}{Line count distance from x-axis. Defaults to 3. Can be negative.}

\item{line2}{Line count distance from y-axis. Defaults to 3. Can be negative.}

\item{cex}{Font size. Defaults to 1.4.}
}
\description{
This function prepares the plotting area with a blank plot
}
\examples{
mtxx(x = expression(bar(x) == sum(frac(x[i], n), i==1, alpha) ~
    (non-sensical ~ axes ~ demo ~ only)),
    y = expression(Log[10] ~ italic(Daphnia ~ pulex) ~ count ~ cm^-3),
    line1=4)
}
\keyword{plot}
