% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{nn}
\alias{nn}
\title{Vectorize and numericize a string of numbers.}
\usage{
nn(x)
}
\arguments{
\item{x}{string or vector of numbers}
}
\description{
This function vectorizes and numericizes a string of numbers. Character elements converted to NA.
}
\examples{
nn(c("a", 1:3, "b", " 1"))
}
\keyword{data}
