% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGrowth.r
\docType{data}
\name{plantGrowth}
\alias{plantGrowth}
\title{Aboveground biomass of imaginary seedlings treated with two levels of N and P in a crossed design.}
\format{
A data frame with 24 rows and 4 variables
}
\usage{
data(plantGrowth)
}
\description{
A sample (fake) dataset showing aboveground biomass of six seedlings per treatment:
}
\details{
\itemize{
  \item N Nitrogen level in watering solution (L = 0.1 g/L; H = 1g/L)
  \item P Phosphorus level in watering solution (L = 0.04; H = 0.4 g/L)
  \item seedSource Blocking factor (Maternal seed source in open pollinated system).
  \item drymass Dry aboveground biomass of each seedling after x weeks.
}
}
\keyword{datasets}
