% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCol.R
\name{plotCol}
\alias{plotCol}
\title{Plots a vecktor from a dataframe}
\usage{
plotCol(x, defaultPalette = FALSE, legend = FALSE, usePlotly = TRUE)
}
\arguments{
\item{x}{a vector from a data.frame to plot}

\item{defaultPalette}{use default color palette or a slightly prettier one,
but limited to 12 colours.}

\item{Legend}{logical with default FALSE. If set to TRUE, a legend will be added to the plot}
}
\value{
ggplot object
#' @importFrom reshape2 melt
}
\description{
Plots a column of a data frame
}
\examples{
plotCol(mydf$grp_tv1)
}

