% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDfCols2.R
\name{plotDfCols2}
\alias{plotDfCols2}
\title{A wraperfunction of plotDfCols}
\usage{
plotDfCols2(df, xColName = "date", defaultPalette = FALSE, facet = FALSE,
  legend = FALSE, usePlotly = TRUE, ...)
}
\arguments{
\item{df}{data.frame to plot}

\item{xColName}{name of column to plot against. Default is 'date', but if
'date' is not present, it defaults to row indexes}

\item{defaultPalette}{use default color palette or a slightly prettier one,
but limited to 12 colours.}

\item{facet}{logical with default FALSE. If set to TRUE, the timeseries are plotted in
seperate plots instead of all in one plot}

\item{legend}{logical with default FALSE. If set to TRUE, the timeseries are plotted with a legend}

\item{...}{passed to ggplot2}
}
\value{
ggplot object
}
\description{
This this a wraperfunction of plotDfCols that exposes the optin to turn of the legend, plotDfCols - plots a data.frame against a specified coulmn
}

