% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.csv.dt.r
\name{read.csv.dt}
\alias{read.csv.dt}
\title{open .csv as data.table object}
\usage{
read.csv.dt(file, datadir = "./data/", sep = ",", header = TRUE)
}
\arguments{
\item{file}{filename No default (required)}

\item{datadir}{Relative (or absolute) path to data directory}

\item{sep}{Use comma for csv}

\item{header}{Defaults to TRUE}
}
\description{
This function imports a csv file into a data.table object
}
\keyword{csv}
