% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeCorrelatedCovariets.R
\name{removeCorrelatedCovariates}
\alias{removeCorrelatedCovariates}
\title{A function to remove correlated covariets}
\usage{
removeCorrelatedCovariates(names, tolerance, myDataFrame)
}
\arguments{
\item{names}{The columnames to be annalyzed}

\item{tolerance}{The highest allowed correlation}

\item{myDataFrame}{Data frame which should be annalyzed}
}
\value{
A data frame with the correlation coefficents in descending order
}
\description{
This function takes column names and looks for correlations between these columns and drops one of them if the correlation is above the tolerance
}
\examples{
names = c('one', 'kpi', 'sun')
removeCorrelatedCovariates(names = names, tolerance = .9, myDataFrame = mydf)
}

