% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeInterceptBehavior.R
\name{removeInterceptBehavior}
\alias{removeInterceptBehavior}
\title{Removes the intercept bahavior of varibles in a decomposition}
\usage{
removeInterceptBehavior(decomp, intercept)
}
\arguments{
\item{decomp}{The decomposition which should be fixed}

\item{intercept}{The intercept column in the decomp as a string}
}
\value{
A decomposition
}
\description{
This function removes the intercept bahavior of all the varibles in a decomposition and returns the new decompsition
}
\examples{
removeInterceptBehavior(removeInterceptBehavior, intercept='allOnes')

}
\author{
Johan Gudmundsson, \email{jgu@blackwoodseven.com}
}

