% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeInterceptBehavior.R
\name{setMinToBase}
\alias{setMinToBase}
\title{Moves the min value of a varible in a decomposition to the intercept}
\usage{
setMinToBase(decomp, baseName, varsToBeShifted, removeZeros = FALSE)
}
\arguments{
\item{decomp}{The decomposition which should be fixed}

\item{baseName}{The intercept column in the decomp as a string}

\item{varsToBeShifted}{The colmuns to be shifted}

\item{removeZeros}{Logic, If true the function will ignore zeros for the columns that are shifted}
}
\value{
A fixed decomposition
}
\description{
This function Sets the min value of a varible in a decomposition to the intercept and removes it from the varible
}
\examples{
setMinToBase(decomp, baseName = 'allOnes', varsToBeShifted = c('trend', 'econmic'))

}
\author{
Johan Gudmundsson, \email{jgu@blackwoodseven.com}
}

