% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smalUtilityFunctions.R
\name{shiftMinToZero}
\alias{shiftMinToZero}
\title{Shifts columns so that the min value for each column in a data frame becomes zero}
\usage{
shiftMinToZero(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame with same structure as the input data frame
}
\description{
Shifts columns so that the min value for each column in a data frame becomes zero, 
if a column is named 'date' will not be affected by the transformation
}
\author{
Johan Gudmundsson, \email{jgu@blackwoodseven.com}
}

