% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceDropbox.r, R/sourceDropbox1.r
\name{sourceDropbox}
\alias{sourceDropbox}
\title{Source from Dropbox file URL}
\usage{
sourceDropbox(dropbox.url)

sourceDropbox(dropbox.url)
}
\arguments{
\item{dropbox.url}{url for shared dropbox link}

\item{x}{Add the x-axis. Defaults to TRUE}

\item{y}{Add the y-axis. Defaults to TRUE}

\item{box}{Add axis lines. Defaults to TRUE}

\item{cex.axis}{Sets axis font size. Defaults to 1.3}
}
\description{
This function loads an .R file from a dropbox link

This function loads an .R file from a dropbox link
}
\examples{
source(sourceDropbox("https://www.dropbox.com/s/gidwa8g23df6uw4/functionlistSample.r?dl=0"))
source(sourceDropbox("https://www.dropbox.com/s/xgbv7suucwf57xq/functionlist.r?dl=0"))
Adds axis in preferred default style

This function prepares the plotting area with a blank plot
plot(1:5, 5:1); axx()
plot(1:5, 5:1); axx(x=FALSE) # Adds y-axis only.
source(sourceDropbox("https://www.dropbox.com/s/gidwa8g23df6uw4/functionlistSample.r?dl=0"))
}
\keyword{dropbox}
\keyword{plot}
\keyword{source}
