% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smalUtilityFunctions.R
\name{subtractMin}
\alias{subtractMin}
\title{Subtract the min value of a vector from each elemnt in the vector}
\usage{
subtractMin(x)
}
\arguments{
\item{x}{vectror}
}
\value{
vector vector with same structure as input vector
}
\description{
Checks if the names of a character vector or colnames of a data frames are valid C varibles
}
\author{
Johan Gudmundsson, \email{jgu@blackwoodseven.com}
}

