% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topCor.R
\name{topCor}
\alias{topCor}
\title{A function to rank correlation coefficients}
\usage{
topCor(df, pairwiseCompleteObs = F, removeDateCol = T)
}
\arguments{
\item{df}{Data frame which should be annalyzed}

\item{pairwiseCompleteObs}{logical, turn on the option use="pairwise.complete.obs" in the calculation of the correlation coefficents}
}
\value{
A data frame with the correlation coefficents in descending order
}
\description{
This function sorts the lower triangular part of the correlation matrix in descending order according to the absolute value of the correlation coefficient
}
\examples{
topCor(df)
}
\author{
Johan Gudmundsson, \email{jgu@blackwoodseven.com}
}

