% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smalUtilityFunctions.R
\name{zerosToNa}
\alias{zerosToNa}
\title{Sets zeros to Na's in vectors and other stuff and returns}
\usage{
zerosToNa(x)
}
\arguments{
\item{x}{vector or matrix, hopfully}
}
\value{
A vector or matrix
}
\description{
This function removes the intercept bahavior of all the varibles in a decomposition and returns the new decompsition
}
\author{
Johan Gudmundsson, \email{jgu@blackwoodseven.com}
}

