% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceDropbox.r
\name{sourceDropbox}
\alias{sourceDropbox}
\title{Source from Dropbox file URL}
\usage{
sourceDropbox(dropbox.url)
}
\arguments{
\item{dropbox.url}{url for shared dropbox link}
}
\description{
This function loads an .R file from a dropbox link
}
\examples{
source(sourceDropbox("https://www.dropbox.com/s/gidwa8g23df6uw4/functionlistSample.r?dl=0"), echo=TRUE)
}
\keyword{dropbox}
\keyword{source}
